Start tracking changes in project
The initial version of this project was completely written, tested, and deemed unfit (this is the code that lives in the `analyze.py` and `generate.py` files in the autoclass module)

A new version was planned, designed, documented (as per the README), and implementation on new version begun (in file `intermediate_representation.py` in the auto_class module)

All without a single change being tracked in git. A serious oversight. Now, moving forward, I plan to track changes consistently